requirements.txt
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==21.4.0
augmax==0.2.2
backcall==0.2.0
bleach==4.1.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.4
docutils==0.18.1
easydict==1.9
einops==0.4.1
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.4.1
huggingface-hub==0.4.0
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
multidict==5.2.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
networkx==2.5.1
nibabel==3.2.2
notebook==6.4.10
numpy==1.19.5
opencv-python==4.6.0.66
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.8.3
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
ptyprocess==0.7.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pyDeprecate==0.3.2
pydicom==2.3.0
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.0
python-dateutil==2.8.2
python-slugify==6.1.2
pytz==2022.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==23.2.0
readme-renderer==34.0
regex==2022.6.2
requests==2.27.1
requests-toolbelt==0.9.1
sacremoses==0.0.53
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
tcga==0.0.22
tensorboardX==2.5.1
terminado==0.12.1
testpath==0.6.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2020.9.3
timm==0.5.4
tokenizers==0.10.3
torch==1.10.2+cu113
torchaudio==0.10.1
torchmetrics==0.8.1
torchvision==0.11.2
tornado==6.0.4
tqdm==4.64.0
traitlets==4.3.3
transformers==4.4.2
twine==2.0.0
typer==0.3.2
typing-extensions==4.1.1
urllib3==1.25.6
uvloop==0.14.0
virtualenv==20.4.3
vit-pytorch==0.35.2
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xgboost==1.5.2
xlrd==1.2.0
XlsxWriter==3.1.2
yarl==1.7.2
zipp==3.1.0
